requirements/local.txt
Django==3.1.6
django-allauth==0.44.0
django-extensions==3.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.3.0
